Remove requiring React
