+++ examples/react/buttons.html
@@ -1,13 +1,13 @@
 <!doctype html>
 <html>
   <head>
-    <link rel="stylesheet" type="text/css" href="../assets/mui/css/mui.css"/>
-    <script src="../assets/mui/react/mui-react.js"></script>
-    
     <!-- for example only -->   
     <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.13.2/react.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.13.2/JSXTransformer.js"></script>
     <!-- end -->
+
+    <link rel="stylesheet" type="text/css" href="../assets/mui/css/mui.css"/>
+    <script src="../assets/mui/react/mui-react.js"></script>
   </head>
   <body>
     <div class="mui-container">
