+++ package.json
@@ -22,9 +22,7 @@
     "streamqueue": "^0.1.3",
     "stringify": "^3.1.0",
     "synthetic-dom-events": "^0.2.2",
-    "vinyl-source-stream": "^1.1.0",
-    "react": "~0.13.3",
-    "classnames": "~2.1.1"
+    "vinyl-source-stream": "^1.1.0"
   },
   "dependencies": {},
   "scripts": {
